:bug: Fix crash options.txt

+++ craftground/craftground/craftground.py
@@ -457,19 +457,22 @@ class CraftGroundEnvironment(gym.Env):
             )
         # Edit options.txt in 1.16
         options_txt = CraftGroundEnvironment.get_env_options_path()
-        with open(options_txt, "r") as file:
-            lines = file.readlines()
-        settings = {}
-        for line in lines:
-            if ":" in line:
-                key, value = line.strip().split(":", 1)
-                settings[key] = value
+        if os.path.exists(options_txt):
+            with open(options_txt, "r") as file:
+                lines = file.readlines()
+            settings = {}
+            for line in lines:
+                if ":" in line:
+                    key, value = line.strip().split(":", 1)
+                    settings[key] = value
 
-        settings["overrideHeight"] = str(self.initial_env.imageSizeY)
-        settings["overrideWidth"] = str(self.initial_env.imageSizeX)
-        with open(options_txt, "w") as file:
-            for key, value in settings.items():
-                file.write(f"{key}:{value}\n")
+            settings["overrideHeight"] = str(self.initial_env.imageSizeY)
+            settings["overrideWidth"] = str(self.initial_env.imageSizeX)
+            with open(options_txt, "w") as file:
+                for key, value in settings.items():
+                    file.write(f"{key}:{value}\n")
+        else:
+            print(f"Options file {options_txt} not found.")
 
         my_env = os.environ.copy()
         my_env["PORT"] = str(port)
